+++ admin/index.php
@@ -382,9 +382,6 @@ class wp_slimstat_admin
         // Add style to the admin menu
         add_action('admin_head', [self::class, 'styling_admin_menu']);
 
-        // Init feedback
-        self::initFeedback();
-
         // Add lock export button in report header
         add_filter('slimstat_report_header_buttons', fn ($_header_buttons, $_report_id) => self::add_lock_export_button($_header_buttons, $_report_id), 10, 2);
 
@@ -2566,55 +2563,6 @@ class wp_slimstat_admin
     }
 
     // END: _create_table
-
-    /**
-     * Init FeedbackBird widget a third-party service to get feedbacks from users
-     *
-     * @url https://feedbackbird.io
-     */
-    private static function initFeedback()
-    {
-        add_action('admin_enqueue_scripts', function () {
-            $screen = get_current_screen();
-
-            if (stristr($screen->id, 'slimview')) {
-                wp_register_script('feedbackbird-widget', 'https://cdn.jsdelivr.net/gh/feedbackbird/assets@master/wp/app.js?uid=01H5FBKA9Z5M2VJWQXZSX4Q7MS', [], null, true);
-                add_filter('script_loader_tag', function ($tag, $handle) {
-                    if ('feedbackbird-widget' === $handle) {
-                        $tag = str_replace('<script ', '<script defer ', $tag);
-                    }
-                    return $tag;
-                }, 10, 2);
-                wp_enqueue_script('feedbackbird-widget');
-                wp_add_inline_script('feedbackbird-widget', sprintf('var feedbackBirdObject = %s;', wp_json_encode([
-                    'user_email' => function_exists('wp_get_current_user') ? esc_attr(wp_get_current_user()->user_email) : '',
-                    'platform'   => 'wordpress-admin',
-                    'config'     => [
-                        'color'         => '#e8294c',
-                        'button'        => __('Feedback', 'wp-sms'),
-                        'subtitle'      => __('Feel free to share your thoughts!', 'wp-sms'),
-                        'opening_style' => 'modal',
-                    ],
-                    'meta' => [
-                        'php_version'    => PHP_VERSION,
-                        'active_plugins' => array_map(fn ($plugin, $pluginPath) => [
-                            'name'    => $plugin['Name'],
-                            'version' => $plugin['Version'],
-                            'status'  => is_plugin_active($pluginPath) ? 'active' : 'deactivate',
-                        ], get_plugins(), array_keys(get_plugins())),
-                    ],
-                ])));
-
-                add_filter('script_loader_tag', function ($tag, $handle, $src) {
-                    if ('feedbackbird-widget' === $handle) {
-                        return preg_replace('/^<script /i', '<script type="module" crossorigin="crossorigin" ', $tag);
-                    }
-
-                    return $tag;
-                }, 10, 3);
-            }
-        });
-    }
 
     public static function get_template($template, $args = [], $return = false)
     {